pig_service/requirements.txt
gTTS
RPi.GPIO
playsound